Tests/LibWeb/Text/expected/canvas/global-composite-operation.txt

normal: Top Left: rgb(125, 60, 190)
normal: Top Right: rgb(230, 55, 85)
normal: Bottom Left: rgb(20, 165, 180)
normal: Bottom Right: rgb(125, 155, 75)
multiply: Top Left: rgb(0, 60, 190)
multiply: Top Right: rgb(205, 55, 35)
multiply: Bottom Left: rgb(0, 115, 180)
multiply: Bottom Right: rgb(120, 100, 30)
screen: Top Left: rgb(125, 125, 255)
screen: Top Right: rgb(230, 110, 95)
screen: Bottom Left: rgb(20, 175, 255)
screen: Bottom Right: rgb(210, 165, 90)
overlay: Top Left: rgb(0, 60, 255)
overlay: Top Right: rgb(230, 55, 45)
overlay: Bottom Left: rgb(0, 165, 255)
overlay: Bottom Right: rgb(190, 145, 40)
darken: Top Left: rgb(0, 60, 190)
darken: Top Right: rgb(205, 55, 45)
darken: Bottom Left: rgb(0, 125, 180)
darken: Bottom Right: rgb(125, 110, 45)
lighten: Top Left: rgb(125, 125, 255)
lighten: Top Right: rgb(230, 110, 85)
lighten: Bottom Left: rgb(20, 165, 255)
lighten: Bottom Right: rgb(205, 155, 75)
color-dodge: Top Left: rgb(0, 125, 255)
color-dodge: Top Right: rgb(230, 110, 70)
color-dodge: Bottom Left: rgb(0, 190, 255)
color-dodge: Bottom Right: rgb(230, 180, 65)
color-burn: Top Left: rgb(0, 60, 255)
color-burn: Top Right: rgb(205, 55, 20)
color-burn: Bottom Left: rgb(0, 110, 255)
color-burn: Bottom Right: rgb(100, 90, 20)
hard-light: Top Left: rgb(125, 60, 250)
hard-light: Top Right: rgb(230, 55, 45)
hard-light: Bottom Left: rgb(0, 165, 235)
hard-light: Bottom Right: rgb(140, 155, 40)
soft-light: Top Left: rgb(0, 95, 255)
soft-light: Top Right: rgb(215, 80, 45)
soft-light: Bottom Left: rgb(0, 140, 255)
soft-light: Bottom Right: rgb(195, 125, 45)
difference: Top Left: rgb(125, 125, 190)
difference: Top Right: rgb(125, 110, 60)
difference: Bottom Left: rgb(20, 100, 200)
difference: Bottom Right: rgb(180, 100, 50)
exclusion: Top Left: rgb(125, 125, 190)
exclusion: Top Right: rgb(125, 110, 85)
exclusion: Bottom Left: rgb(20, 125, 200)
exclusion: Bottom Right: rgb(190, 120, 80)
hue: Top Left: rgb(125, 70, 195)
hue: Top Right: rgb(220, 90, 100)
hue: Bottom Left: rgb(0, 145, 160)
hue: Bottom Right: rgb(120, 150, 70)
saturation: Top Left: rgb(0, 125, 255)
saturation: Top Right: rgb(225, 105, 20)
saturation: Bottom Left: rgb(15, 120, 225)
saturation: Bottom Right: rgb(205, 110, 45)
color: Top Left: rgb(125, 70, 195)
color: Top Right: rgb(230, 85, 100)
color: Bottom Left: rgb(0, 140, 160)
color: Bottom Right: rgb(120, 150, 70)
luminosity: Top Left: rgb(0, 115, 240)
luminosity: Top Right: rgb(185, 90, 25)
luminosity: Bottom Left: rgb(35, 145, 255)
luminosity: Bottom Right: rgb(215, 115, 55)
clear: Top Left: rgb(0, 0, 0)
clear: Top Right: rgb(0, 0, 0)
clear: Bottom Left: rgb(0, 0, 0)
clear: Bottom Right: rgb(0, 0, 0)
copy: Top Left: rgb(0, 0, 0)
copy: Top Right: rgb(0, 0, 0)
copy: Bottom Left: rgb(45, 205, 110)
copy: Bottom Right: rgb(45, 205, 110)
source-over: Top Left: rgb(125, 60, 190)
source-over: Top Right: rgb(230, 55, 85)
source-over: Bottom Left: rgb(20, 165, 180)
source-over: Bottom Right: rgb(125, 155, 75)
destination-over: Top Left: rgb(0, 125, 255)
destination-over: Top Right: rgb(205, 110, 45)
destination-over: Bottom Left: rgb(0, 125, 255)
destination-over: Bottom Right: rgb(205, 110, 45)
source-in: Top Left: rgb(0, 0, 0)
source-in: Top Right: rgb(0, 0, 0)
source-in: Bottom Left: rgb(0, 0, 0)
source-in: Bottom Right: rgb(0, 0, 0)
destination-in: Top Left: rgb(0, 0, 0)
destination-in: Top Right: rgb(0, 0, 0)
destination-in: Bottom Left: rgb(0, 0, 0)
destination-in: Bottom Right: rgb(0, 0, 0)
source-out: Top Left: rgb(0, 0, 0)
source-out: Top Right: rgb(0, 0, 0)
source-out: Bottom Left: rgb(45, 205, 110)
source-out: Bottom Right: rgb(45, 205, 110)
destination-out: Top Left: rgb(0, 0, 0)
destination-out: Top Right: rgb(0, 0, 0)
destination-out: Bottom Left: rgb(0, 0, 0)
destination-out: Bottom Right: rgb(0, 0, 0)
source-atop: Top Left: rgb(0, 0, 0)
source-atop: Top Right: rgb(0, 0, 0)
source-atop: Bottom Left: rgb(0, 0, 0)
source-atop: Bottom Right: rgb(0, 0, 0)
destination-atop: Top Left: rgb(0, 0, 0)
destination-atop: Top Right: rgb(0, 0, 0)
destination-atop: Bottom Left: rgb(45, 205, 110)
destination-atop: Bottom Right: rgb(45, 205, 110)
xor: Top Left: rgb(0, 125, 255)
xor: Top Right: rgb(205, 110, 45)
xor: Bottom Left: rgb(0, 125, 255)
xor: Bottom Right: rgb(205, 110, 45)
lighter: Top Left: rgb(125, 125, 255)
lighter: Top Right: rgb(255, 110, 110)
lighter: Bottom Left: rgb(20, 230, 255)
lighter: Bottom Right: rgb(230, 210, 100)
plus-darker: Top Left: rgb(125, 60, 190)
plus-darker: Top Right: rgb(230, 55, 85)
plus-darker: Bottom Left: rgb(20, 165, 180)
plus-darker: Bottom Right: rgb(125, 155, 75)
plus-lighter: Top Left: rgb(125, 60, 190)
plus-lighter: Top Right: rgb(230, 55, 85)
plus-lighter: Bottom Left: rgb(20, 165, 180)
plus-lighter: Bottom Right: rgb(125, 155, 75)
